Added Decisions and Discussions section
The entire section has been created and much of what needs to be covered is there. A few more things could be added, most notably a introduction. As well, I haven't yet covered the idea of progressing a character by allowing them to choose to reduce their rating in a Motivation if a previous action resulting from that Motivation went poorly for the character.

A few other miscellaneous changes were also made:
- Title formatting
- edits to the Note About Dice