Added dice method for rating generation
Changed the method of random rating generation for main characters to a dice-based method: 4d4 drop highest, subtract 3.

Updated the rules and examples to fit the above.

Moved the general world building distribution to the end and noted why this information is useful. Also tweaked the distribution slightly.

Note: for Future work, I may want to consider adding the following note: "these values are only relevant when the person has had significant training in the skill". But as I'm not sure exactly how I want to do skill development / improvement, I am leaving this out for the moment.