Added Hand-to-Hand combat
Added Hand-to-Hand combat using the rules found in "Combat4".

Playtesting for multi-combatant combat is very thin - I've used a "pairing" / split-the-fight-into-smaller-fights method here, but it may be the case that it is better off just being a free-for all. I just don't really know how to handle movement if it were to be a free-for-all. I'd probably have to drop the footwork contests and do something else.